backend/requirements.txt
Flask
Flask-CORS
Werkzeug
numpy
Pillow
bcrypt
opencv-python
openvino
opencv-python-headless
gdown
